+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.24.3] - 2024-09-09
+
 ### Added
 
 - Add `BiotSavart.autotune` function, which attempts to determine the optimal